+++ ConsoleApp9/Ihlicnan.cs
@@ -8,7 +8,7 @@ namespace ConsoleApp9
 {
     internal class Ihlicnan : Strom
     {
-        public Ihlicnan(int radius1, int vyska1) : base(radius1, vyska1)
+        public Ihlicnan(double radius1, double vyska1) : base(radius1, vyska1)
         {
             radius = radius1;
             vyska = vyska1;
